Moves StringsWatcherC into the cache package.
